Preliminary attempt at creating an additional display. This display is a "Text Only" display. This was done to create a specific result. This outputs an output that could be keyed over live video from an outside source. (such as the ATEM Television Studio)
Would require an additional monitor, (so 4 monitors would need to be used for stage, text only, projection, and control)

Theme implementation is not complete upon this push of the code.

Upon implementation of live video, there should be a live video option for this display.

This is different from the last, as I added some forgotten comments.