Added the option to make the lyrics appear in uppercase on the live view
Many modern worship presentations have their lyrics entirely in uppercase. Whilst I like the look of such presentations, I don't like the thought of haveing the lyrics just in uppercase in the database. With this feature, you can press a new button in the theme-editor, which toggles the uppercase-mode. The lyrics in the database isn't changed at all, it happens purely at rendering