Correct calculation of live/preview display size
There was a subtle(ish) bug in the calculation of the size of the live/preview displays. The ratio of the display was being calculated based on the full size of the split pane canvas that the live/preview display was in. This was then being used to calculate the margins to apply to the canvas to get the actual live/preview being the correct ratio. This was close, but not actually correct when the minimum of 10px margin was taken into account. What should actually be done is subtract the 10px margin from the size of the canvas and then work out what extra margin is needed in addition to the 10px to get a live/preview that is the same ratio as the projection window.