Change GH Action JDK+FX Zulu to just the Temurin JDK (#681)
The Temurin JDK comes preinstalled on the Github-hosted Runners so it saves some time from installing. The FX libraries aren't used for building, but are rather only needed for the runtime, which is taken care of by bundling the standalone installers with the JRE+FX (Zulu) from Azul. (Not needed for build as the JFX libraries are provided by the JavaFX plugin and the dependencies from build.gradle